+++ examples/compute/instance/instance.tf
@@ -25,6 +25,9 @@ variable "ssh_public_key" {
 variable "ssh_private_key" {
 }
 
+variable "kms_key_ocid" {
+}
+
 provider "oci" {
   tenancy_ocid     = var.tenancy_ocid
   user_ocid        = var.user_ocid
@@ -123,6 +126,7 @@ resource "oci_core_instance" "test_instance" {
     # Otherwise, the default boot volume size of the image is used.
     # This should only be specified when source_type is set to "image".
     #boot_volume_size_in_gbs = "60"
+    kms_key_id = var.kms_key_ocid
   }
 
   # Apply the following flag only if you wish to preserve the attached boot volume upon destroying this instance