+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+## 4.107.0 (February 08, 2023)
+
+### Added
+- Support for ADB-S: New ECPU-based SKUs for ADW and ATP
+- Support for Operations Insights : Customizable configuration
+- Support for ADB-D & ADB-CC | Autonomous Data Guard v3
+- Support for ADB-D | Oracle Home Version Control
+- Support for Devops Deploy Helm Attestation with helm args and helm diff
+### Bug Fix
+- Fix oci_core_instance to enable updating KMS Key id associated with the boot volume
+- Resource Discovery is not getting detected for Custom table resource in metering_computation service
+
 ## 4.106.0 (January 31, 2023)
 
 ### Added